Move TapGesture extension to its own file
